🎨 Move all nmm to materials/tools
