+++ requirements.txt
@@ -14,7 +14,7 @@ joblib==1.3.2
 kiwisolver==1.4.5
 llvmlite==0.42.0
 matplotlib==3.8.3
-neutronics-material-maker==0.1.11
+neutronics-material-maker==1.1.4
 nlopt==2.7.1
 meshio==5.3.5
 numba==0.59.1
